requirements.txt
streamlit==1.32.0
langchain==0.1.20
langchain-community==0.0.38
langchain-groq==0.1.3
langchain-text-splitters==0.0.1
chromadb==0.4.24
sentence-transformers==2.7.0
pdfplumber==0.10.3
python-dotenv==1.0.0
tiktoken==0.6.0
